requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
pip==24.3.1
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.5.0
